ordenación de scripts
Algunos archivos estaban mal colocados o duplicados...
