+++ TheFifthElement/TheFifthElement.vcxproj
@@ -130,10 +130,9 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="AnimationEnemyBEUComponent.h" />
-    <ClInclude Include="ColDetectorComponent.h" />
-    <ClInclude Include="ColliderComponent.h" />
-    <ClInclude Include="RedirectTile.h" />
+    <ClInclude Include="src\components\AnimationEnemyBEUComponent.h" />
+    <ClInclude Include="src\components\ColDetectorComponent.h" />
+    <ClInclude Include="src\components\ColliderComponent.h" />
     <ClInclude Include="src\components\AttackBoxComponent.h" />
     <ClInclude Include="src\components\LimitBEU.h" />
     <ClInclude Include="src\Game.h" />
@@ -192,8 +191,8 @@
     <ClInclude Include="src\utils\Vector2D.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="AnimationEnemyBEUComponent.cpp" />
-    <ClCompile Include="ColDetectorComponent.cpp" />
+    <ClCompile Include="src\components\AnimationEnemyBEUComponent.cpp" />
+    <ClCompile Include="src\components\ColDetectorComponent.cpp" />
     <ClCompile Include="src\components\AttackBoxComponent.cpp" />
     <ClCompile Include="src\components\Image.cpp" />
     <ClCompile Include="src\components\LimitBEU.cpp" />
