+++ TheFifthElement/TheFifthElement.vcxproj.filters
@@ -40,10 +40,10 @@
     <ClCompile Include="src\components\CameraComponent.cpp" />
     <ClCompile Include="src\components\AttackBoxComponent.cpp" />
     <ClCompile Include="src\Game.cpp" />
-    <ClCompile Include="ColDetectorComponent.cpp" />
+    <ClCompile Include="src\components\ColDetectorComponent.cpp" />
     <ClCompile Include="src\components\LimitBEU.cpp" />
     <ClCompile Include="src\components\Image.cpp" />
-    <ClCompile Include="AnimationEnemyBEUComponent.cpp" />
+    <ClCompile Include="src\components\AnimationEnemyBEUComponent.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\states\BeatEmUpState.h" />
@@ -101,11 +101,10 @@
     <ClInclude Include="src\utils\Constants.h" />
     <ClInclude Include="src\components\AttackBoxComponent.h" />
     <ClInclude Include="src\Game.h" />
-    <ClInclude Include="RedirectTile.h" />
-    <ClInclude Include="ColliderComponent.h" />
-    <ClInclude Include="ColDetectorComponent.h" />
+    <ClInclude Include="src\components\ColliderComponent.h" />
+    <ClInclude Include="src\components\ColDetectorComponent.h" />
     <ClInclude Include="src\components\LimitBEU.h" />
-    <ClInclude Include="AnimationEnemyBEUComponent.h" />
+    <ClInclude Include="src\components\AnimationEnemyBEUComponent.h" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="TheFifthElement.rc" />
