+++ TheFifthElement/src/components/ColDetectorComponent.cpp
