+++ TheFifthElement/x64/Debug/TheFifthElement.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>D:\GitHub\Selery-Studios\TheFifthElement\x64\Debug\TheFifthElement.exe</FullPath>
+      <FullPath>C:\Selery-Studios\TheFifthElement\x64\Debug\TheFifthElement.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
